Update build instructions for custom Debian image (#9)
Updates the instructions for building custom Debian images outside of
the Nix environment provided by BESSPIN-Environment. A couple of notes:

- The debian-ports archive is now pinned to a past snapshot. Recent
updates to the libc in sid result in incompatibilities with the version
of QEMU provided in Debian 10, causing some of the QEMU-based actions to
fail if not using a snapshot.

- The signing keys for the snapshotted archive have long since expired,
causing signature verification problems when using apt. Those issues can
be side stepped if building the filesystem in a virtual machine, but
just working in a chroot there's not a clean way to turn back time.
While this doesn't make me happy, the best solution that I could find
was to disable signature checks for the snapshot archive.

- When trying to use the original Makefile in bootmem, I ran into
some problem where the system wouldn't fully boot under QEMU, which
interfered with the chainloader process. I haven't tried to diagnose
that problem, and instead have just de-emphasized the use of the
original Makefile in the instructions.